XXE/requirements.txt
boto3
python-docx==0.8.3
